Configured spring security with basic auth. No basic auth configured for index page
