+++ src/test/java/guru/sfg/brewery/web/controllers/BeerControllerTestIT.java
@@ -1,54 +1,13 @@
 package guru.sfg.brewery.web.controllers;
 
-import guru.sfg.brewery.repositories.BeerInventoryRepository;
-import guru.sfg.brewery.repositories.BeerRepository;
-import guru.sfg.brewery.repositories.CustomerRepository;
-import guru.sfg.brewery.services.BeerService;
-import guru.sfg.brewery.services.BreweryService;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.security.test.context.support.WithUserDetails;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-
-import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.httpBasic;
-import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-@WebMvcTest
-public class BeerControllerTestIT {
+public class BeerControllerTestIT extends BaseTestClass{
 
-    @Autowired
-    WebApplicationContext webApplicationContext;
-
-    MockMvc mockMvc;
-
-    @MockBean
-    BeerRepository beerRepository;
-    @MockBean
-    BeerInventoryRepository beerInventoryRepository;
-    @MockBean
-    BreweryService breweryService;
-    @MockBean
-    CustomerRepository customerRepository;
-    @MockBean
-    BeerService beerService;
-
-    @BeforeEach
-    void setUp(){
-        mockMvc = MockMvcBuilders
-                .webAppContextSetup(webApplicationContext)
-                .apply(springSecurity())
-                .build();
-    }
 
     @Test
     @WithMockUser("spring")
