+++ src/test/java/guru/sfg/brewery/web/controllers/IndexControllerTestIT.java
@@ -1,0 +1,15 @@
+package guru.sfg.brewery.web.controllers;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+public class IndexControllerTestIT extends BaseTestClass{
+
+    @Test
+    void testGetIndexSlash() throws Exception {
+        mockMvc.perform(MockMvcRequestBuilders.get("/"))
+                .andExpect(status().isOk());
+    }
+}
